Move PortfolioTokenAsset into own file to clean
